細かいパーツにわける
