+++ src/components/atoms/Button/Base/ButtonSpinner/styles.scss
@@ -1,0 +1,8 @@
+@include themed() {
+  .spinner {
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translateY(-50%) translateX(-50%);
+  }
+}
