+++ src/components/atoms/Button/Base/Inner/styles.scss
@@ -1,0 +1,8 @@
+@include themed() {
+  .inner {
+    display: block;
+    padding: 6px 24px;
+    text-decoration: none;
+    color: inherit;
+  }
+}
