+++ src/components/atoms/Button/Base/styles.scss
@@ -15,13 +15,6 @@
     cursor: pointer;
   }
 
-  .inner {
-    display: block;
-    padding: 6px 24px;
-    text-decoration: none;
-    color: inherit;
-  }
-
   .disabled {
     opacity: .6;
     cursor: default !important;
@@ -29,13 +22,6 @@
 
   .loading {
     color: transparent !important;
-  }
-
-  .loading .spinner {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translateY(-50%) translateX(-50%);
   }
 
   .primary {
